+++ inc/campello_gpu/bind_group.hpp
@@ -28,6 +28,7 @@ namespace systems::leal::campello_gpu {
     class BindGroup {
     private:
         friend class Device;
+        friend class RenderPassEncoder;
         friend class ComputePassEncoder;
         void *native;
         BindGroup(void *pd);