+++ inc/campello_gpu/sampler.hpp
@@ -30,6 +30,8 @@ namespace systems::leal::campello_gpu
     {
     private:
         friend class Device;
+        friend class RenderPassEncoder;
+        friend class ComputePassEncoder;
         void *native;
 
         Sampler(void *pd);